Refactored libunshield CMakeLists to build on windows (#105)
Big thanks Krzysztof for your help!

* Refactored libunshield CMakeLists to build on windows

* Added windows IS to the travis CI

* Left only linux build and testing in travis CI

* Added install prefix to travis script to match the paths currently used by tests

* Added appveyor script for clang and msvc on windows OS

* Added cache with third party dependencies for faster appveoyr rebuilds

* Added upgrade of vcpkg before build

* Added hotfix for temporary msys repo outage

* Added 32 and 64 bit mingw

* Added missing debug/release flags for clang x64

* Fixes for big windows cabfiles

* Changed mingw to gcc compiler name for easier linux integration

* Added linux into appveyor script

* Added exclusions for meaningless windows/linux build matrices

* Added gcc for linux

* Adapting appveyor for linux

* Moved linux build to separate appveyor script

* Removed linux from windows appveyor script

* Renamed default appveyor script to windows specific one

* Restored .travis.yml from original repo

* Added running tests to linux appveyor

+++ lib/internal.h
@@ -78,7 +78,7 @@ void unshield_file_group_destroy(UnshieldFileGroup* self);
 char *unshield_get_base_directory_name(Unshield *unshield);
 long int unshield_get_path_max(Unshield* unshield);
 FILE* unshield_fopen_for_reading(Unshield* unshield, int index, const char* suffix);
-long unshield_fsize(FILE* file);
+long long unshield_fsize(FILE* file);
 bool unshield_read_common_header(uint8_t** buffer, CommonHeader* common);
 
 const char* unshield_get_utf8_string(Header* header, const void* buffer);
